requirements.txt
binance-connector==1.13.0
pytest==7.1.2